I just committed my newest code.  Below is my CVS comment I put in the log. Note the note about calibration at the end. I also put in a check for zero priceElas in subresources (which was causing a problem with nuclear), there aren't any others.
Implemented smooth function for the change in share due to a capacity limit.
Implemented capacity limits for demand sectors.
Turned off calibrations during bracketing (code solves faster and helps calibration).

Added back base-year calibrations for demand sectors in reference.xml.
Added non-energy use oil sub-sector in reference.xml (emits no CO2).
Changed nuclear energy calibration from input to output so that is independent of efficiency assumption.

Because calibrations are turned off during NR solver, code does not calibrate exactly. Seems to be largely a problem for TFE in later years. One way to solve this is to add a call to bisection and then NR (this could use the same derivatives) after the initial solution. It may also be that a call to NR with the same derivatives, but with calibration turned on, would also work. In either case, should be able to do this with only a small increase in model run-time.