First I changed all of the SGM visits from update to start/end.  I also removed any checks on the period in the accept methods for SGM classes to the visitors.
I made expediture into an array so that we can have access to its values for all periods at the end of the model run when the XMLDB output happens.

I changed national accounts from and array of NationalAccount to and array of NationalAccount*. This allowed NationalAccount to reside in RegionSGM instead of Region.

I added and enumToXMLName to expenditure and NationalAccount so that the XML names of the enums are standardized.  I also changed the toDebug in those classes to take advanatage of the new method so each enum print is not hard coded.

I added SGM classes to XMLDBOutputter so that their values can be written to the database. 
I split Region into an abstact base Region and a RegionMiniCAM.  During this split National Accounts got moved into RegionCGE and well as updateMarketplace.

I also changed Input to store a vector of price paid and demand currency. This allows all output to occur at the end of the model run. This required changing several interfaces to include a period argument.

r=jlurz