Output solver diagnostic trace in canonical order.
The vector markets seen by the solver can change between invocations
as the solution-info-filter adds or drops markets from the solution
set.  This causes some markets to jump around in the output, making
it hard to identify markets that seem to be causing solution problems.
This development adds a canonical ordering to the markets in the
calculation.  Solvers that produce diagnostic data will use this
ordering to ensure that the numbering of markets stays consistent in
the output.