Small fixes to configuration file paths and transportation file names.
Corrected an error in the configuration & batch files so the configuration
points to the correct folder for the ssp4_bio_trade.xml file. Also, adjusted
the name of the UCD transportation data file for the core model and set
that as the default file created by the data system instead of SSP1.