solver-diagnostics: create generator function for fxtransform.
Previously fxtransform was just a fixed transformation function.  Now
it's a generator function that takes a solution tolerance as an
argument and returns a transform using that tolerance.  This brings
fxtransform in line with the way all the other transforms work.

Also, delete unused function deltatransform and change all default
values for minimum magnitudes to 1e-3.