Make the preconditioner more conservative in certain cases.
In cases where a resource is being exhausted, the price can become quite
high.  The preconditioner was forcing the trial price to start at the
supply curve upper bound every time, even though that produced demand that
was way too high.  Now the preconditioner is a little more conservative in
choosing trial prices in these circumstances.