Increase the `guess_max` parameter in read_csv to 10M lines.
This is sort of a proxy for "the whole file".  It turns out that
read_csv won't let you specify an arbitrarily large value for this
parameter.  It tops out at some random-looking number north of 20M.  I
went with the 10M figure as a nice, round number that should be
sufficiently large.

This changes fixes a couple of bugs related to read_csv mistakenly
detecting a column as integer if all of the entries up to `guess_max`
look like integers.  Then, if it sees a floating-point number
somewhere down the line, it chokes.  Scanning far enough that you can
be sure of seeing the floating point values is the simplest
workaround.