Changes logits after 02-14-2018 meeting in which we discussed how to clean up the code for consistency in both the energy and ag uses.
In particular we avoid using the term "cost" and instead use "value"
since in the energy system the logit functions will operate on costs and
in the ag it will be profit rates.

No longer choose node profit rates to be the implied/intrinsic value and
instead just use the profit rate of the dominant child in the nest.
This simplifies the code and makes it so that we no longer have to cary
and "output value" and "base value".  We also no longer need to even
calculate the implied/intrinsic value any more.

Add in some error checking to ensure logits have the correct sign
depending on where it is used.  Also ensure that "base value" is never
negative as this would result in sharing working in the wrong direction
(increased profit rate leads to lower share for instance).